# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nestling API
  version: "1"
  description: |
    The HTTP API of the Nestling server, for generating client code. `docs/API.md` explains the
    same API in prose; keep both in step with `src/routes/` (`tests/openapi.rs` checks that every
    route is listed here).

    - **Auth:** `Authorization: Bearer <token>`. `EventSource` (which can't set headers) may pass
      `?access_token=<token>` instead.
    - **Times in requests:** `"now"`, a local time in the family's time zone
      (`"2026-10-08T14:30"`) or RFC 3339 with an offset. An omitted `start` means now.
    - **Times in responses:** RFC 3339; family data in the family's time zone, account and token
      data in UTC.
    - **Units:** metric only (`_ml`, `_g`, `_cm`, `_c` = °C, `_seconds`).
    - **IDs:** opaque strings (UUIDv7). Where a client may supply its own id it must be a UUID.
    - Things you don't have access to answer `404`, like things that don't exist.
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
  - url: https://nestling.example.com/api/v1
security:
  - bearer: []
...
paths:
...
  /children/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [children]
      operationId: getChild
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Child" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [children]
      operationId: updateChild
      summary: Send only what changes; `null` clears `birth_date` / `sex`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                birth_date: { type: string, format: date, nullable: true }
                sex: { allOf: [{ $ref: "#/components/schemas/Sex" }], nullable: true }
                medicines:
                  type: array
                  description: Replaces the whole list.
                  items: { $ref: "#/components/schemas/Medicine" }
                reminders:
                  type: array
                  description: Replaces the whole list (one per type).
                  items: { $ref: "#/components/schemas/Reminder" }
                book:
                  allOf: [{ $ref: "#/components/schemas/Book" }]
                  description: Replaces the whole object.
      responses:
        "200":
          description: The updated child
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Child" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [children]
      operationId: deleteChild
      summary: Delete the child and its records
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Doesn't exist, or you don't have access
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, forbidden, not_found, conflict, too_many_requests, upstream_error, internal]
            message: { type: string }
      example: { error: { code: bad_request, message: "a diaper must be wet, dirty or dry" } }
...
    Sex:
      type: string
      enum: [female, male, other]
...
    Child:
      type: object
      required: [id, family_id, name, birth_date, sex, created_at, updated_at, photo_version, medicines, reminders, book]
      properties:
        id: { type: string }
        family_id: { type: string }
        name: { type: string }
        birth_date: { type: string, format: date, nullable: true }
        sex: { allOf: [{ $ref: "#/components/schemas/Sex" }], nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        photo_version:
          type: integer
          format: int64
          nullable: true
          description: Changes whenever the photo does; null without a photo.
        medicines:
          type: array
          items: { $ref: "#/components/schemas/Medicine" }
        reminders:
          type: array
          items: { $ref: "#/components/schemas/Reminder" }
        book: { $ref: "#/components/schemas/Book" }
    Book:
      type: object
      description: >-
        The baby book's own pages as short texts by key (the app uses e.g. `birth_time`,
        `birth_place`, `hair`, `eyes`, `full_name`, `name_why`, `world_songs`, `price_coffee`).
        Keys: 1–40 lowercase letters, digits or `_`; at most 80 of them; values are strings of at
        most 4,000 characters. Empty strings and nulls are dropped.
      additionalProperties: { type: string, maxLength: 4000 }
    Medicine:
      type: object
      description: |
        A medicine schedule. Its doses are the child's `health` events with `kind: medicine` and the
        same name (case and extra spaces ignored).
      required: [name, every_hours]
      properties:
        name: { type: string, maxLength: 80 }
        every_hours: { type: number, minimum: 0.5, maximum: 168 }
        max_per_day: { type: integer, minimum: 1, maximum: 24, description: At most this many doses in any 24 hours }
        dose: { type: number, minimum: 0, description: The usual dose (fills the app's form) }
        dose_unit: { type: string }
        remind: { type: boolean, default: false, description: Notify phones when the next dose is due (server with notifications) }
    Reminder:
      type: object
      description: |
        Shown by the app when due; also sent to phones by a server with notifications set up.
        `sleep` counts time awake.
      required: [type, after_minutes]
      properties:
        type: { type: string, enum: [feed, sleep, diaper, pump] }
        after_minutes: { type: integer, minimum: 15, maximum: 1440 }
...