# src/routes/children.rs
use axum::{
    extract::{Path, State},
    http::StatusCode,
    Json,
};
use chrono::NaiveDate;
use chrono_tz::Tz;
use serde::Deserialize;
use serde_json::{json, Value};

use crate::{
    auth::{child_access, child_book_access, family_tz, member_role, require_member, AuthUser, BOOK_VIEWER},
    error::{bad, ApiJson, AppResult},
    schedule::{check_medicines, check_reminders, parse_medicines, parse_reminders, Medicine, Reminder},
    state::AppState,
    util::{fmt_time, new_id, now_ms},
};

type ChildRow = (String, String, String, Option<String>, Option<String>, i64, i64, Option<i64>, String, String, String);

fn child_value(row: ChildRow, tz: Tz) -> Value {
    let (id, family_id, name, birth_date, sex, created_at, updated_at, photo_version, medicines, reminders, book) = row;
    json!({
        "id": id, "family_id": family_id, "name": name, "birth_date": birth_date, "sex": sex,
        "created_at": fmt_time(created_at, tz), "updated_at": fmt_time(updated_at, tz),
        // Changes whenever the photo does (null: no photo); fetch it from /children/{id}/photo.
        "photo_version": photo_version,
        "medicines": parse_medicines(&medicines),
        "reminders": parse_reminders(&reminders),
        "book": parse_book(&book),
    })
}

/// A child as a book viewer sees it: without the medicine schedules and reminders.
pub fn book_view(child: &mut Value) {
    if let Some(obj) = child.as_object_mut() {
        obj.remove("medicines");
        obj.remove("reminders");
    }
}

const CHILD_COLS: &str = "id, family_id, name, birth_date, sex, created_at, updated_at, \
    (SELECT p.updated_at FROM child_photos p WHERE p.child_id = children.id) AS photo_version, medicines, reminders, book";

pub async fn children_json(state: &AppState, family_id: &str, tz: Tz) -> AppResult<Vec<Value>> {
    let rows: Vec<ChildRow> = sqlx::query_as(&format!("SELECT {CHILD_COLS} FROM children WHERE family_id = ? ORDER BY created_at"))
        .bind(family_id)
        .fetch_all(&state.db)
        .await?;
    Ok(rows.into_iter().map(|r| child_value(r, tz)).collect())
}

pub async fn child_json(state: &AppState, child_id: &str, tz: Tz) -> AppResult<Value> {
    let row: ChildRow = sqlx::query_as(&format!("SELECT {CHILD_COLS} FROM children WHERE id = ?"))
        .bind(child_id)
        .fetch_one(&state.db)
        .await?;
    Ok(child_value(row, tz))
}

/// The baby book's pages: short texts by key (`birth_place`, `name_why`, `world_songs`…). The
/// keys are the app's business; the server only keeps the object small and flat.
pub type Book = serde_json::Map<String, Value>;

pub fn parse_book(text: &str) -> Value {
    match serde_json::from_str::<Value>(text) {
        Ok(v @ Value::Object(_)) => v,
        _ => json!({}),
    }
}

/// Keys of lowercase letters, digits and `_` (40 at most), string values of 4,000 characters at
/// most, 80 keys at most. Empty strings and nulls are dropped.
pub fn check_book(book: &mut Book) -> AppResult<()> {
    book.retain(|_, v| !(v.is_null() || v.as_str().is_some_and(|s| s.trim().is_empty())));
    if book.len() > 80 {
        return bad("book has too many entries (80 at most)");
    }
    for (k, v) in book.iter_mut() {
        if k.is_empty() || k.len() > 40 || !k.bytes().all(|b| b.is_ascii_lowercase() || b.is_ascii_digit() || b == b'_') {
            return bad(format!("book key '{k}' must be 1-40 lowercase letters, digits or _"));
        }
        let Some(text) = v.as_str() else {
            return bad(format!("book.{k} must be a string"));
        };
        if text.chars().count() > 4000 {
            return bad(format!("book.{k} is too long (4000 characters at most)"));
        }
        *v = Value::String(text.trim().to_string());
    }
    Ok(())
}

fn check_sex(sex: &Option<String>) -> AppResult<()> {
    match sex.as_deref() {
        None | Some("female") | Some("male") | Some("other") => Ok(()),
        Some(_) => bad("sex must be 'female', 'male' or 'other'"),
    }
}

/// Insert a child; used by the API and the Nara importer.
pub async fn insert_child(state: &AppState, family_id: &str, name: &str, birth_date: Option<NaiveDate>, sex: Option<String>) -> AppResult<String> {
    let id = new_id();
    let now = now_ms();
    sqlx::query("INSERT INTO children (id, family_id, name, birth_date, sex, created_at, updated_at) VALUES (?, ?, ?, ?, ?, ?, ?)")
        .bind(&id)
        .bind(family_id)
        .bind(name)
        .bind(birth_date.map(|d| d.to_string()))
        .bind(sex)
        .bind(now)
        .bind(now)
        .execute(&state.db)
        .await?;
    Ok(id)
}

pub async fn list(State(state): State<AppState>, user: AuthUser, Path(family_id): Path<String>) -> AppResult<Json<Value>> {
    let role = member_role(&state.db, &family_id, &user.id).await?;
    let tz = family_tz(&state.db, &family_id).await?;
    let mut children = children_json(&state, &family_id, tz).await?;
    if role == BOOK_VIEWER {
        children.iter_mut().for_each(book_view);
    }
    Ok(Json(json!({ "children": children })))
}

#[derive(Deserialize)]
pub struct CreateChildReq {
    name: String,
    #[serde(default)]
    birth_date: Option<NaiveDate>,
    #[serde(default)]
    sex: Option<String>,
}

pub async fn create(State(state): State<AppState>, user: AuthUser, Path(family_id): Path<String>, ApiJson(req): ApiJson<CreateChildReq>) -> AppResult<(StatusCode, Json<Value>)> {
    require_member(&state.db, &family_id, &user.id).await?;
    if req.name.trim().is_empty() {
        return bad("name is required");
    }
    // Required for new babies; existing ones (and Nara imports) may still have none.
    if req.birth_date.is_none() {
        return bad("birth_date is required");
    }
    check_sex(&req.sex)?;
    let id = insert_child(&state, &family_id, req.name.trim(), req.birth_date, req.sex).await?;
    let tz = family_tz(&state.db, &family_id).await?;
    let child = child_json(&state, &id, tz).await?;
    state.publish(&family_id, "child", "created", child.clone());
    Ok((StatusCode::CREATED, Json(child)))
}

pub async fn get(State(state): State<AppState>, user: AuthUser, Path(child_id): Path<String>) -> AppResult<Json<Value>> {
    let ctx = child_book_access(&state.db, &child_id, &user.id).await?;
    let mut child = child_json(&state, &child_id, ctx.tz).await?;
    if ctx.book_only {
        book_view(&mut child);
    }
    Ok(Json(child))
}

/// `birth_date` / `sex` can be cleared by sending null.
#[derive(Deserialize)]
pub struct UpdateChildReq {
    name: Option<String>,
    #[serde(default, deserialize_with = "double_option")]
    birth_date: Option<Option<NaiveDate>>,
    #[serde(default, deserialize_with = "double_option")]
    sex: Option<Option<String>>,
    /// Replaces the whole list.
    #[serde(default)]
    medicines: Option<Vec<Medicine>>,
    #[serde(default)]
    reminders: Option<Vec<Reminder>>,
    /// The baby book's pages; replaces the whole object.
    #[serde(default)]
    book: Option<Book>,
}

fn double_option<'de, T, D>(d: D) -> Result<Option<Option<T>>, D::Error>
where
    T: serde::Deserialize<'de>,
    D: serde::Deserializer<'de>,
{
    Option::<T>::deserialize(d).map(Some)
}

pub async fn update(State(state): State<AppState>, user: AuthUser, Path(child_id): Path<String>, ApiJson(req): ApiJson<UpdateChildReq>) -> AppResult<Json<Value>> {
    let ctx = child_access(&state.db, &child_id, &user.id).await?;
    if let Some(name) = &req.name {
        if name.trim().is_empty() {
            return bad("name cannot be empty");
        }
        sqlx::query("UPDATE children SET name = ? WHERE id = ?").bind(name.trim()).bind(&child_id).execute(&state.db).await?;
    }
    if let Some(bd) = req.birth_date {
        sqlx::query("UPDATE children SET birth_date = ? WHERE id = ?")
            .bind(bd.map(|d| d.to_string()))
            .bind(&child_id)
            .execute(&state.db)
            .await?;
    }
    if let Some(sex) = req.sex {
        check_sex(&sex)?;
        sqlx::query("UPDATE children SET sex = ? WHERE id = ?").bind(sex).bind(&child_id).execute(&state.db).await?;
    }
    if let Some(mut list) = req.medicines {
        check_medicines(&mut list)?;
        sqlx::query("UPDATE children SET medicines = ? WHERE id = ?")
            .bind(serde_json::to_string(&list)?)
            .bind(&child_id)
            .execute(&state.db)
            .await?;
    }
    if let Some(list) = req.reminders {
        check_reminders(&list)?;
        sqlx::query("UPDATE children SET reminders = ? WHERE id = ?")
            .bind(serde_json::to_string(&list)?)
            .bind(&child_id)
            .execute(&state.db)
            .await?;
    }
    if let Some(mut book) = req.book {
        check_book(&mut book)?;
        sqlx::query("UPDATE children SET book = ? WHERE id = ?")
            .bind(serde_json::to_string(&book)?)
            .bind(&child_id)
            .execute(&state.db)
            .await?;
    }
    sqlx::query("UPDATE children SET updated_at = ? WHERE id = ?").bind(now_ms()).bind(&child_id).execute(&state.db).await?;
    let child = child_json(&state, &child_id, ctx.tz).await?;
    state.publish(&ctx.family_id, "child", "updated", child.clone());
    Ok(Json(child))
}

pub async fn delete(State(state): State<AppState>, user: AuthUser, Path(child_id): Path<String>) -> AppResult<StatusCode> {
    let ctx = child_access(&state.db, &child_id, &user.id).await?;
    sqlx::query("DELETE FROM children WHERE id = ?").bind(&child_id).execute(&state.db).await?;
    state.publish(&ctx.family_id, "child", "deleted", json!({ "id": child_id }));
    Ok(StatusCode::NO_CONTENT)
}

/// Largest profile picture accepted (the app sends a ~512 px square, far smaller).
pub const PHOTO_LIMIT: usize = 5 * 1024 * 1024;

/// Image type from the file's first bytes (the Content-Type header isn't trusted).
pub fn image_type(data: &[u8]) -> Option<&'static str> {
    match data {
        [0xFF, 0xD8, 0xFF, ..] => Some("image/jpeg"),
        [0x89, b'P', b'N', b'G', ..] => Some("image/png"),
        [b'R', b'I', b'F', b'F', _, _, _, _, b'W', b'E', b'B', b'P', ..] => Some("image/webp"),
        _ => None,
    }
}

/// Set the profile picture: the raw image (JPEG, PNG or WebP) as the request body.
pub async fn put_photo(State(state): State<AppState>, user: AuthUser, Path(child_id): Path<String>, body: axum::body::Bytes) -> AppResult<Json<Value>> {
    let ctx = child_access(&state.db, &child_id, &user.id).await?;
    let Some(content_type) = image_type(&body) else {
        return bad("the photo must be a JPEG, PNG or WebP image");
    };
    let now = now_ms();
    sqlx::query(
        "INSERT INTO child_photos (child_id, content_type, data, updated_at) VALUES (?, ?, ?, ?)
         ON CONFLICT(child_id) DO UPDATE SET content_type = excluded.content_type, data = excluded.data, updated_at = excluded.updated_at",
    )
    .bind(&child_id)
    .bind(content_type)
    .bind(body.as_ref())
    .bind(now)
    .execute(&state.db)
    .await?;
    let json = child_json(&state, &child_id, ctx.tz).await?;
    state.publish(&ctx.family_id, "child", "updated", json.clone());
    Ok(Json(json))
}

pub async fn get_photo(State(state): State<AppState>, user: AuthUser, Path(child_id): Path<String>) -> AppResult<axum::response::Response> {
    use axum::{http::header, response::IntoResponse};
    child_book_access(&state.db, &child_id, &user.id).await?;
    let row: Option<(String, Vec<u8>)> = sqlx::query_as("SELECT content_type, data FROM child_photos WHERE child_id = ?")
        .bind(&child_id)
        .fetch_optional(&state.db)
        .await?;
    let Some((content_type, data)) = row else {
        return Err(crate::error::AppError::NotFound("photo"));
    };
    Ok(([(header::CONTENT_TYPE, content_type), (header::CACHE_CONTROL, "private, no-cache".to_string())], data).into_response())
}

pub async fn delete_photo(State(state): State<AppState>, user: AuthUser, Path(child_id): Path<String>) -> AppResult<StatusCode> {
    let ctx = child_access(&state.db, &child_id, &user.id).await?;
    sqlx::query("DELETE FROM child_photos WHERE child_id = ?").bind(&child_id).execute(&state.db).await?;
    let json = child_json(&state, &child_id, ctx.tz).await?;
    state.publish(&ctx.family_id, "child", "updated", json);
    Ok(StatusCode::NO_CONTENT)
}

#[cfg(test)]
mod tests {
    use super::image_type;

    #[test]
    fn sniffs_images() {
        assert_eq!(image_type(&[0xFF, 0xD8, 0xFF, 0xE0]), Some("image/jpeg"));
        assert_eq!(image_type(b"\x89PNG\r\n"), Some("image/png"));
        assert_eq!(image_type(b"RIFF\0\0\0\0WEBPVP8 "), Some("image/webp"));
        assert_eq!(image_type(b"<svg></svg>"), None);
    }
}